+++ lab7/SlideLib/Shapes/Shape.cs
@@ -1,5 +1,5 @@
 ﻿using RGBAColor = System.UInt32;
-using RectD = SlideLib.Rect<double>;
+using RectD = SlideLib.Rect<int>;
 
 namespace SlideLib.Shapes
 {
